Warn about modifying instance in hooks
I ended up spending a fair amount of time tracking down unpredictable behavior as a result of doing exactly that. I ended up writing components to display the "updated" count just for visual feedback on how often updates were taking place. I then wrote a component to do so and found that it worked when written certain ways but not in others, and yielded unpredictable results. I tried tracking them down.

I had thought about the circular reference/infinite loop but presumed that Vue had detection logic and broke the chain when it detected the looping (and it seemed to). So I was trying to figure out what I didn't understand about components when one didn't work as I expected. See https://stackoverflow.com/questions/41780872/vue-js-component-updates-inconsistently for details if interested.

It might make more sense not to introduce instance lifecycle events this early in the guide, but that's more major change.