Fix Lodash not loading from unpkg
For some reason, unpkg is not loading lodash.min.js, which is breaking the question example for watchers on the computed properties page. I've checked the .min links for axios and vue on unpkg, and it looks like it is just an issue with lodash. For example https://unpkg.com/lodash@4.13.1/lodash.min.js will return an error saying

> Cannot find file "/lodash.min.js" in package lodash@4.13.1

Going to https://unpkg.com/lodash@4.17.4/lodash.min.js results in the same error but with the 4.17.4 tag.

I have checked the directories on unpkg and they are listed as being there. Someone may want to get in contact with unpkg to let them know this, as for now I've suggested to use cdnjs for the lodash@4.17.4 min.js file which fixes the error the computed properties page was getting (_ is not defined).