Update comparison between `data` and `vm` objects
The comparison `vm.a === data.a` does not show that `vm` and `data` reference the same object, but rather that the `a` property of both objects is the same value. My updated example compares the `vm` and `data` object directly, thereby adding to the next two examples which indicate that changes to the `vm` object's properties are reflected in the `data` (and vice-versa).