Add note in list.md on usage of v-for with v-if
Noticed that this was mentioned in the guide, but also mentioned in
category "Essential" in the style guide as a pattern to avoid. Therefore
suggesting this change where we warn that this is not best-practice and
should be avoided.

If you can add the label "hacktoberfest" to this PR before reviewing, I
would be grateful.