Improvement Suggestion
Having UI status as boolean is prone to bugs. In this case, it's correct, but in other cases, the user can end up with `error` and `loading` turned on due to bad code, causing a bug. By having status as a string and not boolean, we can be safe against this type of bug.