Remove smoothScroll from example
Since smoothScroll is undefined, it gives an error in the console when trying to execute the example: example1.items.push({ message: 'Baz' })

This might confuse new people. And because it does not do anything because of it being undefined, it perhaps is better not to have it in the code at all.