Introduce performance regression CI check
Using [Gimbal](https://github.com/ModusCreateOrg/gimbal) and GitHub Actions to ensure Performance Budgeting on VueJS.org homepage.

This CI action will
* [x] ensure lighthouse tests don’t fall below thresholds
* [x] ensure a11y rules are observed
* [x] ensure heap memory doesn’t go over the threshold
* [x] automatically create a PR comment with items that are failing

The thresholds are loose to reflect the current state of the built site. Some of the tests are skipped as they are taken care of on the server side post-deployment.