Update components.md
The example of "Using v-model on Components" was a little unclear for me. As a newcommer at Vue I couldn't understand what is "searchText" and where it comes. I had to investigate it in jsfiddle and it took some time to clarify. 
We can make this example much easier to understand with this small addition.

```
new Vue({
    el: '#app',
    data: {
        searchText: "text"
    }
})
```