Add tip about `done` callbacks in transition hooks
It took me *hours* to figure out why my `onAfterEnter` function was not being called. It turns out that if your `onEnter` function *accepts* the `done` parameter but does not *use* the parameter then the `onAfterEnter` function never gets called.

JSFiddle to show the behaviour: https://jsfiddle.net/adamalton/7f0kca4m/54/

I considered making a bug report for it, but from looking at discussions on similar issues it seems that Vue's sniffing of the function parameters is intentional, so the behaviour is arguably deliberate. So instead I'm just adding a tip to the docs so that others after me don't need to go through the same pain 🙂.