Use consistent object key syntax
Forgive me if I'm missing something (I'm just learning Vue from this tutorial now) but it seems like there is an unnecessary inconsistency between the key syntax for the `active` and `text-danger` classes. Is there a reason why `text-danger` is rendered with single quotes while `active` is not? If so, it should be stated. But it seems to me like an inconsistency that could confuse a novice like myself, or at least make it slightly less reader friendly.