Wrong last name algorithm
This is a classic error when trying to get the last name on a non-English language. E.g. in Latin America is common to have two last names.
It would be  better if the expression gets simplified to `this.lastName = names[1]` assuming just one first name and one last name