Add comment to components code example
In the code example for registering a new component I've added a comment that the registration must take place prior to creating the root Vue instance. This is mentioned later in the Component Registration section but as a new user I spent a lot of time trying to figure out why my component wasn't rendering when trying to build a a small web app with the information I had learned from the introduction/essentials sections because I didn't realize order was important. In hindsight it makes sense why that is necessary but I still think it may be useful to point this out.