Update syntax.md
Using "attributeName" in chrome will cause error: Property or method "attributeName" is not defined on the instance but referenced during render. Make sure that this property is reactive, either in the data option, or for class-based components, by initializing the property.
For the new folks to Vuejs, please update it.