Update syntax.md
Giving examples using camelcase for dynamic attribute names is confusing to beginners, as this won't work when they try to use it in an HTML template. Having the explanation at the end of that section (line 170) is good, but the examples given should lead learners in the right direction whenever possible.

If people have strong feelings against snake case, then no special casing could be used instead, ie. attributename.