Update unit-testing-vue-components.md
Previous text under **When to avoid this pattern** said nothing about avoiding this pattern. It sounded like a disclaimer, so that's what I called it and put it at the top. I thought at least a word should be said about when to avoid tests though, so I wrote it at the bottom.